+++ root_modules/com_gae_user_staff_group/v1_0_0/staticfiles/js/list/list.controller.js
@@ -52,28 +52,37 @@
       if (action != 1) {
         return false;
       }
-      GAEUI.pageLoading().play();
-      vm.bulkDelete = "";
-      if (vm.selectedDeleteId.length === 0) {
-        GAEUI.pageLoading().stop();
-        GAEUI.notification().playError('Please select some staff group');
-      } else {
-        var dataSend = {
-          "ids": vm.selectedDeleteId
-        };
-        CUR_MODULE.apiPost('start/bulk_delete', dataSend).then(function (res) {
+      GAEUI.confirmBox().play("Are you sure?","please confirm your action.",function(bool){
+        if(bool){
+          GAEUI.confirmBox().stop();
+          GAEUI.pageLoading().play();
+          vm.bulkDelete = "";
+          if (vm.selectedDeleteId.length === 0) {
+            GAEUI.pageLoading().stop();
+            GAEUI.notification().playError('Please select some staff group');
+          } else {
+            var dataSend = {
+              "ids": vm.selectedDeleteId
+            };
+            CUR_MODULE.apiPost('start/bulk_delete', dataSend).then(function (res) {
+              vm.selectedDeleteId = [];
+              vm.bulkDelete = "";
+              if (res.ok) {
+                fetchListing();
+                GAEUI.pageLoading().stop();
+                GAEUI.notification().playComplete("Delete staff group complete");
+              } else {
+                GAEUI.pageLoading().stop();
+                GAEUI.notification().playError('Can not delete staff group');
+              }
+            });
+          }
+        } else {
           vm.selectedDeleteId = [];
           vm.bulkDelete = "";
-          if (res.ok) {
-            fetchListing();
-            GAEUI.pageLoading().stop();
-            GAEUI.notification().playComplete("Delete staff group complete");
-          } else {
-            GAEUI.pageLoading().stop();
-            GAEUI.notification().playError('Can not delete staff group');
-          }
-        });
-      }
+        }
+        GAEUI.confirmBox().stop();
+      },"Done");
     }
 
     function onClickSort(keyname) {
